+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Changed
+- Improved `get_icons_for_user` function icon name parsing
+
 ## [1.6.0] - 2018-09-13
 ### Added
 * Allow overriding plugins page admin menu removal from user meta with meta key `_airhelper_admin_show_plugins`